Feat: schema 공통 부분 분리
