+++ src/schemas/schedules.schema.ts
@@ -1,17 +1,5 @@
 import { z } from "@/utils/ko-zod";
 
-export const storeIdParamsSchema = z.object({
-  params: z.object({
-    storeId: z.string().length(8),
-  }),
-});
-
-export const scheduleIdParamsSchema = z.object({
-  params: z.object({
-    scheduleId: z.string().length(16),
-  }),
-});
-
 export const createScheduleBodySchema = z.object({
   body: z.object({
     userId: z.string().length(12),
